Asny call of getting the data.
It is possible to write a route resolver which will load data before the component is even initalised meaning the component dosn't need to load the data as it's in a route object.

The problem is the route now tells the component how it must collect the data but it means other components may get access to this data without reloading it.